requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
atomicwrites==1.3.0       # via pytest
attrs==19.3.0             # via pytest
certifi==2019.9.11        # via requests
chardet==3.0.4            # via requests
idna==2.8                 # via requests
importlib-metadata==0.23  # via pluggy, pytest
more-itertools==7.2.0     # via pytest, zipp
packaging==19.2           # via pytest
pathlib2==2.3.5           # via pytest
pluggy==0.13.0            # via pytest
py==1.8.0                 # via pytest
pyaes==1.6.1              # via telethon
pyasn1==0.4.7             # via rsa
pyparsing==2.4.2          # via packaging
pytest==5.2.2
requests==2.22.0
rsa==4.0                  # via telethon
six==1.12.0               # via packaging, pathlib2
telethon==1.33.0
urllib3==1.25.6           # via requests
wcwidth==0.1.7            # via pytest
zipp==0.6.0               # via importlib-metadata

aiogram~=3.14.0
python-dotenv~=1.0.1